Added coarse hashing
Coarse hashing is currently unused for repost detection, but they may be used in the near future. This change will allow existing installs to build a database of both coarse hashes and fine hashes (the default for repost detection), so that coarse hashes can immediately be used once they're implemented in the repost detection.